Support new clang name mangling of main to __main_argv_argv
This change just ensure that the linker preserves main whichever
name it is known by.

See the corresponding binaryen and llvm changes:

https://github.com/WebAssembly/binaryen/pull/2671
https://reviews.llvm.org/D75277
