+++ system/lib/standalone/standalone.c
