+++ system/lib/standalone/standalone_wasm_stdio.c
