+++ tests/core/test_hello_argc.c
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2016 The Emscripten Authors.  All rights reserved.
+ * Emscripten is available under two separate licenses, the MIT license and the
+ * University of Illinois/NCSA Open Source License.  Both these licenses can be
+ * found in the LICENSE file.
+ */
+
+#include <stdio.h>
+int main(int argc, char** argv) {
+  printf("hello, world! %d\n", argc);
+  return 0;
+}
