+++ tests/core/test_hello_argc.out
@@ -1,0 +1,1 @@
+hello, world! 1
