+++ tests/other/metadce/mem_O3_STANDALONE_WASM.funcs
@@ -1,4 +1,4 @@
-$__original_main
+$__main_void
 $_start
 $dlmalloc
 $main
